Add dedicated mock classes instead of duplicating code.
